Package request model/repo/controller added

+++ src/main/java/com/chinabox/delivery/controllers/ApplicationController.java
@@ -1,14 +1,21 @@
 package com.chinabox.delivery.controllers;
 
 import com.chinabox.delivery.dao.AddressRepository;
+import com.chinabox.delivery.dao.AuthTokenRepository;
 import com.chinabox.delivery.dao.UserRepository;
+import com.chinabox.delivery.model.AuthToken;
 import com.chinabox.delivery.model.User;
 import com.chinabox.delivery.model.UserAddress;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Scope;
+import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.context.WebApplicationContext;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,6 +23,10 @@ import java.util.Optional;
 @RestController
 @RequestMapping("rest/user")
 public class ApplicationController {
+
+
+    private @Autowired
+    HttpServletRequest request;
 
     @Autowired
     @Resource
@@ -26,7 +37,17 @@ public class ApplicationController {
     AddressRepository addressRepository;
 
     @Autowired
-    User user;
+    @Resource
+    AuthTokenRepository authTokenRepository;
+
+
+    @Bean
+    @Scope(value = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
+    public User requestUser() {
+        AuthToken authToken = this.authTokenRepository.getByKey(request.getHeader("Authorization"));
+        return authToken.getUser();
+
+    }
 
     @PostMapping(value = "register")
     public void createUser(@RequestBody User user) {
@@ -43,9 +64,15 @@ public class ApplicationController {
 
     @PostMapping(value = "setAddress")
     public void setUserAddress(@RequestBody UserAddress address) {
-        address.setUser(this.user);
+        address.setUser(requestUser());
+        if (addressRepository.findByUser(requestUser()) != null) {
+            for (UserAddress a : addressRepository.findByUser(requestUser())) {
+                addressRepository.delete(a);
+            }
+        }
         addressRepository.save(address);
     }
+
 
     @GetMapping(value = "findById")
     public Optional<User> findById(Long id) {
@@ -54,7 +81,12 @@ public class ApplicationController {
 
     @DeleteMapping(value = "delete")
     public void deleteUserById(Long id) {
-        userRepository.delete(userRepository.findById(id).get());
+
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.delete(userRepository.findById(id).get());
+        } else {
+            System.out.println("User does not exist");
+        }
 
     }
